make SimpleWriter a handler and remove WriteHandler
Keep an alias for backward compatibility.
