+++ tools/pyosmium-get-changes
@@ -33,7 +33,7 @@ from osmium.replication import server as rserv
 from osmium.replication import newest_change_from_file
 from osmium.replication.utils import get_replication_header
 from osmium.version import pyosmium_release
-from osmium import SimpleHandler, WriteHandler
+from osmium import SimpleHandler, SimpleWriter
 
 
 import re
@@ -228,9 +228,9 @@ def main(args):
 
         log.debug("Starting download at ID %d (max %d MB)" % (startseq, options.outsize))
         if options.outformat is not None:
-            outhandler = WriteHandler(options.outfile, 4096*1024, options.outformat)
+            outhandler = SimpleWriter(options.outfile, filetype=options.outformat)
         else:
-            outhandler = WriteHandler(options.outfile)
+            outhandler = SimpleWriter(options.outfile)
 
         endseq = svr.apply_diffs(outhandler, startseq, max_size=options.outsize*1024,
                                  simplify=options.simplify)
